stt/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
websockets>=13.1
numpy>=1.26.0
librosa>=0.10.2
soundfile>=0.12.1
pydub>=0.25.1
openai-whisper>=20231117
torch>=2.4.0
python-multipart>=0.0.9
pydantic>=2.9.2
python-dotenv>=1.0.1
prometheus-client>=0.20.0
